Rename Vulkan wrapper classes to XJ prefix, remove XJVulkanAddModelAssets
- Rename VulkanInstance->XJVulkanInstance, VulkanPhysicalDevices->XJVulkanPhysicalDevices
- Rename VulkanQueue->XJVulkanQueue, VulkanSurface->XJVulkanSurface, VulkanImageView->XJVulkanImageView
- Unify all Platform classes under XJ naming convention
- Remove XJVulkanAddModelAssets (unused legacy code)
- Update all includes and Platform CMakeLists accordingly
- Refine VulkanCommon.h and various engine classes

+++ Platform/Private/Edit/XJGlfwWindow.cpp
@@ -164,7 +164,7 @@ namespace XJ
             auto *kWindow = static_cast<XJGlfwWindow *>(glfwGetWindowUserPointer(window));
             if(kWindow)
             {
-                XJMouseMoveEvent mouseMovedEvent{ static_cast<float>(xpos), static_cast<float>(ypos) };
+                XJMouseMoveEvent mouseMovedEvent{ xpos, ypos };
                 XJEventDispatcher::XJGetInstance()->DispatchEvent(mouseMovedEvent);
             }
         });
